[R5D4]: Refactor to use a content variable
A content variable was used to store the content of HTML element
and then string interpolation was used to include the content in
the JSX element.

A class name variable was created; called myClassName. This variable
served as the class and then string interpolation was used to include
the variable in the JSX sript.